fix: make completion big decimal

+++ src/main/java/org/ntnu/idi/idatt2106/sparesti/sparestibackend/model/Challenge.java
@@ -39,5 +39,5 @@ public class Challenge {
     @Column(nullable = false, name = "TYPE")
     private ChallengeType type;
 
-    @Transient private double completion;
+    @Transient private BigDecimal completion;
 }
